* Display an error message in the chat when channel change or shop entry fails.  * Set up communication infrastructure for shop server.  * Game server supports migrating player to the shop server.  * Implemented the necessary channel scan, chatroom, buddy notifications, party notifications, and guild notifications functionality between game and shop server.  * Partially implemented ShopHandler so that a player connecting to the server will at least see the cash shop interface.

+++ trunk/src/argonms/common/net/external/CommonPackets.java
@@ -354,7 +354,7 @@ public final class CommonPackets {
 		//for wedding or "love" rings?
 		lew.writeShort((short) 0);
 		lew.writeShort((short) 0);
-		lew.writeShort((byte) rings.size());
+		lew.writeShort((short) rings.size());
 		for (Ring ring : rings) {
 			lew.writeInt(ring.getPartnerCharId());
 			lew.writePaddedAsciiString(Player.getNameFromId(ring.getPartnerCharId()), 13);
